screens

+++ FGC_T3/RawDataModel.cpp
@@ -170,7 +170,7 @@ bool RawDataModel::createBackFaceTexture()
         glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_T, GL_REPEAT);
         glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST);
         glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
-        glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA16F, MainData::rootWindow->getSize().x, MainData::rootWindow->getSize().y, 0, GL_RGBA, GL_FLOAT,
+        glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA32F, MainData::rootWindow->getSize().x, MainData::rootWindow->getSize().y, 0, GL_RGBA, GL_FLOAT,
                      NULL);
     } catch (...) {
         return false;
